Works now. I've got it more stable with shorter spring rest length, tighter steps and more dampening. In example below there are 10 points laid out along x-axis every 100 feet. Spring rest length is 1 foot. You can see the contraction in the screenshot below. You can see from the forces in the m_Velocity field and the delta between m_oldPosition and m_Position that the point is moving to the left. What i'm not sure, because the dynamics are not working yet is how well the dampening works over time and does acceleration stay under control or not.
Regarding dynamics - I'm still experimenting with timers and having trouble having the stopwatch trigger an update when the interval is hit. I have recreated the old dynTimer and am using a stopwatch but the only way I can hit KeepTime() is in the Evaluate loop for dynTimer. I want to send a pulse downstream every time the stopwatch hits a specified interval in milliseconds. I feel a mental block coming on here - I need some Stephen magic or at least a hint from Stephen or Ian.

Todo on the dynDynamicRelaxation:

parse input point and curve lists, convert to particles and springs. Right now it just uses a stub setup method in the constructor
Make the output points and curves be handled non-destructively, this will speed things up a lot
Manage changes to input lists (add,remove,change)
Expose stepsize as an input port.
Coloring of springs would fun but adding OSTs is a PIA in revit, maybe use AVF instead.