Setup MIConvexHullPlugin.dll to work with Dynamo
Needed to create and apply a strong name (a crypto hash key) in MIConvexHullPlugin project to allow registering this dll in the GAC and then referencing the assembly in Dynamo.
This is to get around the failure we were seeing trying to reference this pre-compiled dll from dyanmo\packages directory.
see http://support.microsoft.com/kb/315682

DynBench.LoadBuiltinTypes() - I wrapped this in a try-catch as we were having problems with an
external dll (MIConvexHullPlugin.dll) not loading correctly from \dynamo\packages
because the dll did not have a strong name by default and was not loaded into the GAC.
The exceptions are now caught but if there is an exception no built-in types are loaded.