dynDynamicRelaxation and dynWatch3d updates
I just updated both nodes to simply pass through the particle system. If
watch3d gets a container it casts it as a PS and then extracts the
particles and springs. It might be doing a little bit of double work
still as it is only drawing lines and not points but at least there are
not intermediate XYZs and curves. Is there a way to draw lines and
points in Helix? i did not see in a cursory look.

also updated dynParticleSystem a bit to take more inputs. New example
added as well.