Dynamo can connect to ArduMote via dynUDPListener!
This is an idea that Tom V wants to explore. We were talking about doing
this through arduino connections like this
http://www.youtube.com/watch?v=rxpmoFbpgd0 but because it's just sending
UDP packets I figured it would be easy enough to just make a UDP
listener node in Dynamo and cut out the middle-man.
see new samples also, more info on ArduMote here.
https://itunes.apple.com/us/app/ardumote/id449100779?mt=8