Fix to trap settings shutdown causing UI test failure
During shutdown, the dynSettings.Controller is set to null

The even to close the SearchView is fired, which causes the view to try and execute a search. As the controller is now null, this causes a crash.

Currently all the UI tests fail due to this. When this is committed, they pass or are inconclusive.