Fixes Null marshaling in collection and List tests
- CollectionMarshaler was throwing exception when any item was null
while getting the approx DS type of the items in the collection. Fixed
this by proper null check.
- Removed ArbitraryDimensionArrayImport from IList parameters as by
default IList will be marshaled as arbitrary dimension array.
- ArbitraryDimensionArrayImport attribute is applied to item parameter
of AddItemToFront method, so that a collection can be added to the front
of the list.
- indices parameter of method List.RemoveItemAtIndex is changed to int[]
from object so that only integers are passed. VM will automatically
promote single object to a collection while calling this method.