Change hand off execution notification to the end
Right now upstream notification occurs at the beginning. This results in the vis. manager update message being sent before the transaction is committed, and is causing issues with the new tighter concurrency bounds.

Given that each call to this assumes that all of the work for the notification has been done, it should happen after this method is complete, not before.