Revert commit to fix MAGN-3456
I had removed the static reference to the view, to ensure we didn't have
static memory problems. But doing so revived the old behavior that you
could start two sessions of Dynamo on Revit.

We need to investigate what APIs are available from Revit to disable the
Dynamo button so we can remove this static reference once and for all.