added new methods for generating labels from ID and exposed this in dynamo for testing
added methods for transforming list of curves from initial position to space of specific surface by ID

fixed bug in projected center of surface as a polygon that was returning the center of the polygon non projected - this method needs some more work, would be better to find the interior of a surface somehow, slow but robust is to cast a ray from this projected point and walk inside the polygon/surface, similar to magic wand alto.

fixed a bug in transformation tracking of surfaces, now we add other surfaces that have been folded coplanar with the parent from other branches to the parent, so they are not lost, BUT we don't build a transformation entry with them until we actually transform the parent.  May still need to be added for disconnected sets that overlap and are moved away, will need to find a good Test/Defect for this.