Fixing UpdateManagerUITests failures
CheckForProductUpdate() assumed that UpdateManager.Instance will always
be of type UpdateManager, but the UpdateManagerUITests mock
IUpdateManager and set the instance for testing.
So this method failed in casting.

Exposed Configuration property to return IUpdateManagerConfiguration
from IUpdateManager interface so that this method doesn't assume
anything.

Updated testcases to mock the IUpdateManager with configuration
settings.