Fix for selection box dragged outside the canvas
Removed the mouse leave event on Workspace view.
Captured the mouse event on LeftButtonDown and  released the mouse
capture on LeftButtonUp. This will ensure that selection is not lost
even when the mouse is dragged outside the canvas.
Handled the mouse lost focus event on DragCanvas. This will ensure that
selection is removed if the mouse is released outside the canvas.