Integrate RuntimeCore to the VM
1. Integrate RuntimeCore instance to the runtime. The runtime vm will
refer to this when accessing runtime properties.
2. RuntimeCore is passed in alongside Core in most functions, this is to
enable moving runtime properties to RuntimeCore in smaller chunks
without having to break the function calls. Eventually, all instances of
Core will be removed from the runtime
3. Instantiate DebugProperties in RuntimeCore and let the VM refer to
only this during a debug session
4. The Codegen still generates debug information during compilation and
stores in its own DebugProperties structure. Only the Codegen can refer
to this
5. After compilation is complete, the Codegen DebugProperties is copied
to the RuntimeCore