Parse doubles during migration as culture invariant.
There are people who live in other countries. In some of those
countries, they use a comma as the decimal separator. Strange as it
might seem, it is required that we treat all doubles as having
culture-invariant formatting or shit will be in the wrong place. WHY CAN
WE NOT GET THIS RIGHT?!!!!