Introduce Notications object.
Introduce the Notifications object and the NotificationsControl. These
replace the previous mechanism for display run-time messages on Dynamo's
lower bar.

The Notifications object is a singleton with one public method,
PostNotification, which in turn raises the NotificationPosted event.

The NotificationsControl is a UserControl which binds to a
NotificationsViewModel as its DataContext. It listens for
NotificationPosted events. When an event is handled the view model sets
its CurrentNotification which then updates the UI. A small amount of
code-behind is added to the NotificationsControl to allow the control to
handle the DynamoView's PreviewMouseDown event, which then call's a
method on the ViewModel to clear the current notification.