add showDesignOptionsDialogCommand to the list of commands that toggles when selection is changed this is set in DynamoViewModel.cs
add a new menu item in the edit menu that calls the correct command to bring up the dialog

hook this command to fire an event requesting the dialog and attach the handler that brings up this dialog in the DynamoView xaml cs file, this follows the same structure of the CustomNodeProperty Command and Dialog window.

Like NewCustomNodeCommand we'll also need to create a new command that actually builds the design options object on the workspace, current plan is to just serialize this object like any other graph element on save so all input parameter states travel simply with the .dyn file.