Remove static Instance of UpdateManager
- UpdateManager class is no more a singleton class, hence static
property "Instance" is removed.
- DynamoModel holds IUpdateManager instance, which is initialized via
StartConfiguration parameter. If StartConfiguration doesn't have a valid
instance of IUpdateManager then it creates a instance of UpdateManager
with default configuration parameter.
- Clients can set IUpdateManager while initializing DynamoModel, thru
StartConfiguration parameter.
- A static method GetProductVersion() is introduced to return product
version, so that this can be called even when DynamoModel is not
initialized.
- A new property IsUpdateAvailable is defined to return true only if
update info is downloaded and we have a valid update or force update is
set true.
- The green cloud on UI is bound to this property, so the cloud should
be visible only after data has been downloaded.
- Tests has been updated to make use of UpdateManager and some new tests
are added to test IsUpdateAvailable property.