Fix preview upstream...
We were too fancy with the toggling of visibility. When IsUpstreamVisible is toggled we should just clear and re-render. There are too many things that could have happened while the node was not previewing upstream to guarantee that just flipping visibilities back on will look correct.

This also adds a better extension methods for finding upstream visible nodes.