Fix all command and visibility bindings...
Moving 3 commands to the HelixWatch3DViewModel caused binding errors for the context menus which expected these command to be in the DynamoViewModel.
Watch3DView's context menu bindings are simplified by setting the tag of the outer grid to contain the view model for the watch 3d view. This allows the context menu items to binding directly to that view model instead of using PlacementTarget.