Remove the ICleanup interface...
The ICleanup interface was made a long time ago and used by the visualization manager. It is no longer required as the methods for clearing workspaces during shutdown have improved and are the proper way to request clearing. Here, the ICleanup interface is removed, and the references to its event, CleaningUp, are removed from the IDynamoModel interface.