Remove the lost focus handler...
This handler was originally added to attempt to solve an issue with the
user getting locked into 3D mode when they clicked on a second screen when
the nav key was down. This was not an effective solution because this handler
gets fired constantly, whenever we return focus to search for example. I will
validate the MAGN-5055 holds after this removal.