Fixed: when input the fraction without unit, the output will stil be displayed correctly in feet. Problem: 2 problems 1. When user input a fraction without unit, the output returned is always 0. 2. When user input a fraction with unit of feet, the output returned is always 0.
Reaons:
Problem 1: the function SetValueFromString() in Units.cs only try to parse input of a whole number. Thus the fraction input without unit will not return a correct result (in this case 0 is returned).
Problem 2: String Pattern in the function ParseLengthFromString() in units.cs only account for whole number or decimal number of input of unit(feet). No faction input of unit(feet) is handle.