An additional basic autocompletion provider
I propose adding a secondary generic autocomplete provider that pops up automatically after every 2nd character and provides suggestions from  the built-in python methods and the most common "words" used in the script up to this point. That effectively helps to greatly reduce the chance for spelling mistakes and can speed up writing.