Avoid rendering Gizmo multiple times on rotate
A change of behavior in helix when upgrading from v2015 is that the
CameraChanged event gets fired several times, once for each individual
property of Camera that changed.

In order to avoid rendering the Gizmo several times, which leads to
choppy camera rotation, we limit the rendering of the Gizmo to only be
done when the position of the camera changed.