Make sure extension window closes with Dynamo (#10557)
View extensions that add a Window to the extension bar, run the risk of
leaving a window open when Dynamo closes, resulting in a zombie process
. The extension could avoid the problem in either of these two ways:

1. Setting the Owner property of the Window they create to the one
obtained from ViewLoadedParam.DynamoWindow.

2. Calling the Close method on the Window they created when the
Shutdown life-time method is called on the extension.

However, Dynamo should not rely on view extensions doing this. In order
to be able to still close no matter what, we assign the Owner property
explicitly in DynamoView.AddExtensionTabItem.