Handle memory outage on render (cherry pick) (#10569)
This brings the fix from master to the helix-upgrade branch. The only
difference is that the test now mocks the AggregateRenderPackage
function directly, because the function mocked in master no longer
exists. The visibility was changed to internal virtual in order to be
able to mock it, also using InternalsVisibleTo.