Register custom node before package load reset
This fixes a problem where existing custom nodes in the home workspace
became unresolved after a package that contained binaries was loaded.

The cause of the problem was that the compiled function was not
available at the time of the VM reset. Now the data is registered on
package load, by queueing it in pendingCustomNodeSyncData. This results
in a CompileCustomNodeAsyncTask being scheduled before the update of
the home workspace graph takes place.