Support heterogeneous arrays in repl FEP matching (#10639)
The no longer failing test pictures this scenario better than words.
Before this change, we would not be able to find a function endpoint
that matched the arguments, even with replication and loose type check.

This was caused by a heuristic done at the Replicator that I have named
'array sampling'. Basically, when getting type examples to produce the
reduced parameters for replication, this optimization would sample only
the first non-empty array it found in the given parent array layer.

In the failing test, the first non-empty array led to reduced params of
types string and int for the first argument of Dictionary.ValueAtKey.
As the function was expecting a dictionary, and no cast is possible
from those types, the FEP was not resolved.

In order to solve this problem, an alternative replication parameter
production strategy that does not sample arrays was introduced. By
applying this in the failing test case, we are now also inspecting the
nested array that contains the dictionary, which leads to resolving the
FEP as compliant.

For performance reasons, this strategy is used only in the context of
the most permissive FEP matching, which is case 6, as the amount of
reduced parameters to consider for replication could increase greatly.