Executive fepRun mismatch on graph update
Under circumstances related to multiple node deletion, the current
executive of the runtime core can present a mismatch on the value of
the fepRun field. This is used to indicate that an execution is part of
a function endpoint call.

This mismatch is circumvented by following a pattern already being used
in BuiltInFunctionEndpoint and other places, where an interpreter is
created, resulting in a cloned executive but with the fepRun using a
value provided in the interpreter constructor.

There is also another problem fixed here related to the built-in
transpose function not behaving correctly when provided values that are
a mix of empty arrays and scalar values.