[DYN-2680] Write logs to std out instead of creating new files for each test. (#10651)
* Write logs to std out instead of creating new files for each test.

This is to avoid storing these large number of logs on the VM which is affecting its performance over time.

* Fix for API break change.

* Fixing the failing tests.

These tests verify that the logs are written to dynamo console and to a file. When we changed the behaviour to write the logs to std out in test mode, there were failing. Fixed them accordingly.

* Avoid the stream writer object, no longer needed.