Fix transpose bug and handle executive revert (#10656)
Under circumstances related to multiple node deletion in the graph
provided, the current executive of the runtime core can present a
mismatch on the value of the fepRun field. This later caused a crash
when the VM, assuming that it was calling a function endpoint, tried
to get a procedure node at an invalid index.

The cause of the this anomaly was an incorrect error handling at
JILFunctionEndPoint.Execute, where an error in a downstream call would
cause the revert of the replaced executive never to happen.

Also, the function that caused the improperly handled error is fixed.
Now the built-in transpose function should behave correctly when
provided values that are a mix of empty arrays and scalar values.