Make Helix render more backwards compatible (#10707)
The fixed directional light coming from top is eliminated. This light
behaved completely different with the previous shader. Basically, this
was being moved with the view so there was no light coming from above,
but this was actually coming from the camera, which is what the head
light now does.

Also, some color calculations are changed in the shader so that the
final color outcome looks how it used to.