Bind Python engine selection in cm to model
The content menu for Python nodes and Python string nodes was not
binding to the 'Engine' property of the model, but was instead updating
the check state of the menu items manually. This caused the check state
to get out of sync with the model when the engine selected using the
Python node 'Edit' dialog.