Loggins Models Test Methods - Code Smells
All this changes were reported in Sonarqube as code smells (4 code smell entries).

In the DynamoAnalyticsClientTest.cs file I removed the instance using the IsUsageReportingApproved property (the coverage was not changed).

In the DynamoModelCommandsTest.cs file for the public void BeginDuplicateConnectionTest() method I replaced the node Watch by a DSFunction.

In the DynamoModelEventsTest.cs I re-wrote the TestOnDeletionStarted() test method to not use the Annotations property (deprecated).