Add support for alternate resource names
The alternate name supported is the same as the original resource name,
but adds the culture name before the extension.

This is needed to support the workflow the localization team is using
for translations. The tool they are using produces resources with names
that do not match the name of the original embedded resource.

Example: German translation of NoContent.html will be searched in de-DE
satellite assembly by the names of:
1. NoContent.html
2. NoContent.de-DE.html