Add trace back for pythonnet errors (#10782)
When errors occur in Python.NET, a PythonException is thrown. This
class, unfortunately, does not put the trace back information in the
message. It is available in the stack trace, but showing that did not
seem a good idea. Instead, reflection is used to access the private
field that holds the trace back. The trace back, if available, is
appended to the message of a new exception that wraps the original.