DYN-2561 - Coverage Connectors folder
I added the ConnectorModelConnectTest() method to increase the coverage in the ConnectorModel class.

The  internal class InvalidPortException was deleted due that is no used (no reference was found in all the Dynamo solution). I got approval from Aaron and Michael in the Autodesk Slack Channel.

For the DeserializeCore(XmlElement nodeElement, SaveContext context) I deleted the commented code. I got approval from Aaron and Michael in the Autodesk Slack Channel.