Prevent SO for unknown type in CPython (#10845)
When unmarhsalling output data from a Python script using the CPython
engine, a StackOverlowException would occur for certain uncommon types.
The unmarshaller is trying to recursively convert the output by calling
'AsManagedObject', but in cases where Python.NET does not have a
specialized representation, like it is the case for 'weakref', it just
returns a different object pointing to the same handle. This is
causing the unmarshaller to call himself again with basically the same
underlying Python object, triggering the stack overflow.

The problem is fixed by explicitly checking if the returned object by
'AsManagedObject' has the same handle as the original one, in which
case an exception is thrown with a proper message.