Convert Python iterable to List in function calls
Adds support to convert any Python iterable to a list in the context of
a function call using the CPython engine. The support is added by
relaxing the check in the custom decoder to allow any iterable to come
through, not only sequences.

This should now allow converting dictionary view types, set and
frozenset.