Convert Python iterable to IList and IEnumerable (#10879)
Adds support to convert any Python iterable to an IList or IEnumerable
in the context of a function call using the CPython engine. The decoder
now checks if the PyObject returned is iterable instead of a sequence,
which is supported by the function we are using in Python.NET. Also,
IEnumerable, both generic and non-generic, is declared as a target type
to allow more conversion scenarios.

This should now allow converting dictionary view types, set and
frozenset to IList. It should also allow these and any other types that
are iterable (list, tuple, dictionary) to be converted to IEnumerable.