Dummy Python evaluator for missing engine (#10897)
When building the AST for a Python node an exception is no longer
thrown when the engine is missing. Instead the class and method of a
dummy evaluator are returned. This evaluator simply returns an error
message indicating that the configured engine is missing.

Doing this allows to recover from the absence of a selected Python
engine, as evaluation errors are retried on the next run. Also, since
this is now an evaluation error, it is displayed over the node as a
warning.