Increase code coverage of the NodeModel class (#10913)
* Added tests for NodeModel

* Added tests for NodeModel

* Revert "Added tests for NodeModel"

This reverts commit 4434d5fcbd1f452ac17d9555d73c6a185a37d5c8.

* Update NodeModelTests.cs

* Removed a test and marked the properties as obsolete

* Changed Obsolete for IsObsolete applying it only to the setter

* Updated message

* Added tests for NodeModel

* Deprecated PrintExpression and removed the associated test.

* DYN-2561 - Coverage Connectors folder

I added the ConnectorModelConnectTest() method to increase the coverage in the ConnectorModel class.

The  internal class InvalidPortException was deleted due that is no used (no reference was found in all the Dynamo solution). I got approval from Aaron and Michael in the Autodesk Slack Channel.

For the DeserializeCore(XmlElement nodeElement, SaveContext context) I deleted the commented code. I got approval from Aaron and Michael in the Autodesk Slack Channel.

* Added tests for Attributes

* Replaced a string with its associated resource.

* Added some more tests for NodeModel

Co-authored-by: Roberto Tellez <roberto.tellez@globant.com>