Load library in test, required by extension (#10919)
The Python migration extension now depends on PythonNodeModels.
However, the dependency is not private, so the dll might be absent
when running tests in parallel. Since the extension is installed but a
dependency is missing, activating it will result in load exceptions.
That is the case for the test here.

The problem is fixed by forcing the dll to be loaded indirectly, by
adding DSCPython as a library to preload. That makes sure
PythonNodeModels will be available during the test run and that the
extension won't fail because of its absence.