Add LogLevel.ConsoleOnly for DynamoLogger (#10940)
This log level allows to log messages only to the console, skipping
file logging. The addition of a new level valuie was required because
counterintuitively LogLevel.Console logs to both the console and file.

This new level is currently used only for logging with DynamoPrint, but
the overload was added as a public function in DynamoLogger, so it can
be used in arbitrary code, like ZT nodes.