Fix whitespace during Python 2to3 migration
The 2to3 migration script, by design, does not take care of normalizing
whitespaces. Python 3, as opposed to Python 2, enforces that spaces and
tabs cannot be mixed. In order to get a working script, a reindentation
is performed over the migrated content. This is only done in case the
code uses tabs, otherwise there is no need to do changes to indent.