Fix line number for IronPython errors
The IronPython evaluator was adding an 'import sys' and a new line as a
prefix to the code entered in a Python script node. This was causing
errors to be reported in the the line next to where they actually
occurred.

The issue is fixed by removing the prefixed code, since it was not
really needed.