Fix line number for IronPython errors (#10987)
The IronPython evaluator was adding an 'import sys' and a new line as a
prefix to the code entered in a Python script node. This was causing
errors to be reported in the the line next to where they actually
occurred.

The issue is fixed by removing the prefixed code, since it was not
really needed. However, the import is still done in order to remain
backwards compatible with graphs that may be using 'sys' without
importing it. This is done without modifying the user's source.