Define default Python engine by system and user
Adds the ability to define a default Python engine for the system. This
is done through a new property in the model start configuration, named
'DefaultPythonEngine'.

Also adds the ability for users to define through a setting their own
default. This can be set to either engine or to fallback to the system
default, whatever it is.

Serialization is tweaked so that 'IronPython2' remains the
deserialization default but is not the absolute default. Engine will be
null on node creation but become loaded before added to the graph, to
avoid errors on automatic execution. For this, a new function was added
to nodes named 'OnBeforeAdd' which is called before the node is added
to the graph. More importantly, this function makes the DynamoModel
available to nodes, making it possible to get both the system default
and user preferred engine.