Obsolete Python evaluation events
Python evaluation events are exposing internals of the engines they
use, which is not a good practice. Meanwhile we decide the future of
these events in the context of moving out Python engines from Core, we
will declare these events as 'Obsolete' to make clear our intentions of
making changes in the near future.