Move Python print function to local scope (#11007)
This fixes build issues that appeared after merging 'print' function
changes for CPython3. Those changes moved the print function to the
globalScope which caused the 'DynamoPrintLogsToConsole' test to fail.
The reason for this is that Python unit tests, which don't use a model,
still initialize the globalScope but do not provide a logger, making
later logging tests fail.