Define default Python engine by system and user (#10998)
Adds the ability to define a default Python engine for the system. This
is done through a new property in the model start configuration, named
'DefaultPythonEngine'.

Also adds the ability for users to define through a setting their own
default. This can be set to either engine or to fallback to the system
default, whatever it is.

Serialization is tweaked so that 'IronPython2' remains the
deserialization default but is not the absolute default. Engine will be
unspecified on node creation but become loaded as soon as the property
is accessed, to avoid errors on automatic execution.

In order to make system and user settings available on the node, and
avoiding further changes to the API, the properties were made static
and internal.