Dyn 3026 code coverage analysis folder (#10989)
* DYN-3026-CodeCoverage-AnalysisFolder

I created several test methods for increasing the code coverage in the next classes (Analysis project).
- Label
- SurfaceData
- AnalysisExtensions
- Utils

Also I removed two methods from the SurfaceData.cs file since both have cero references and both are private.

* DYN-3026-CodeCoverage-AnalysisFolder

Fix for make the Analysis assembly visible to the AnalysisTests project.

* DYN-3026-CodeCoverage-AnalysisFolder-CodeReview

Changes asked by Aparajit.

* DYN-3026-CodeCoverage-AnalysisFolder-CodeReview2

I re-wrote the test for the Label class since Michael asked me to use a Label inside a Code Block node, then I created 3 dyn files with one CodeBlock creating a Label, one compiling and other two with errors (on purpose) when passing parameters to the Label.ByPointAndString method.

Also by Michael response I undid the changes done by Aaron in the PR https://github.com/DynamoDS/Dynamo/pull/8189.

Finally I applied some changes that the developer Ian Keough did like 5 years ago (HelixRenderPackage.cs -> CleanTag) but were discarded in the PR below, with the code as it is now in the Dynamo repo  when creating a Label inside a Code Block it shows a strange symbol in the Background Preview. The only way to show the label text correctly was using the syntax "label:text", then Ian's discarded change was re-applied by me and now we don't need to use the weird Label syntax.
https://github.com/DynamoDS/Dynamo/pull/4637

* DYN-3026-CodeCoverage-AnalysisFolder-CodeReview2

I applied some changes that the developer Ian Keough did like 5 years ago (HelixRenderPackage.cs -> CleanTag) but were discarded in the PR below, with the code as it is now in the Dynamo repo when creating a Label inside a Code Block it shows a strange symbol in the Background Preview. The only way to show the label text correctly was using the syntax "label:text", then Ian's discarded change was re-applied by me and now we don't need to use the weird Label syntax.
https://github.com/DynamoDS/Dynamo/pull/4637

* DYN-3026-CodeCoverage-AnalysisFolder-CodeReview2

I moved the LabelTests to the DynamoCoreWpfTests project inside the Libraries folder.
I removed several dll in the AnalysisTests project that are producing erros in the mono.sln.